requirements.txt
Flask==2.3.3
SQLAlchemy==1.4.50
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
python-dotenv==1.0.0
alembic==1.12.1
pytest==7.4.3
gunicorn==21.2.0